@@ -1,7 +1,7 @@
 import Combine
 import Foundation
 
-public protocol MeetingNoUserJoinedUseCaseProtocol {
+public protocol MeetingNoUserJoinedUseCaseProtocol: Sendable {
     var monitor: AnyPublisher<Void, Never> { get }
     func start(timerDuration duration: TimeInterval, chatId: HandleEntity)
 }
