@@ -1,14 +1,19 @@
 import Combine
 import MEGAChatSdk
 import MEGADomain
+import MEGASwift
 
-public final class MeetingNoUserJoinedRepository: NSObject, MeetingNoUserJoinedRepositoryProtocol {
-    public static var sharedRepo = MeetingNoUserJoinedRepository(chatSDK: .sharedChatSdk)
+public final class MeetingNoUserJoinedRepository: NSObject, MeetingNoUserJoinedRepositoryProtocol, @unchecked Sendable {
+    public static let sharedRepo = MeetingNoUserJoinedRepository(chatSDK: .sharedChatSdk)
     
+    @Atomic
     private var subscription: AnyCancellable?
         
     private let chatSDK: MEGAChatSdk
-    private(set) var chatId: HandleEntity?
+    
+    @Atomic
+    private var chatId: HandleEntity?
+    
     private let source = PassthroughSubject<Void, Never>()
         
     public var monitor: AnyPublisher<Void, Never> {
@@ -26,24 +31,27 @@ public final class MeetingNoUserJoinedRepository: NSObject, MeetingNoUserJoinedR
     }
 
     public func start(timerDuration duration: TimeInterval, chatId: HandleEntity) {
-        self.chatId = chatId
-        subscription = Just(Void.self)
+        self.$chatId.mutate { $0 = chatId }
+        let subscription = Just(Void.self)
             .delay(for: .seconds(duration), scheduler: DispatchQueue.global())
             .sink { _ in
                 self.source.send()
                 self.cleanUp()
             }
+        $subscription.mutate {
+            $0 = subscription
+        }
     }
     
     private func cleanUp() {
-        subscription?.cancel()
-        subscription = nil
-        chatId = nil
+        $subscription.wrappedValue?.cancel()
+        $subscription.mutate { $0 = nil }
+        $chatId.mutate { $0 = nil }
     }
 }
 
 extension MeetingNoUserJoinedRepository: MEGAChatCallDelegate {
-    public func onChatCallUpdate(_ api: MEGAChatSdk, call: MEGAChatCall) {
+    nonisolated public func onChatCallUpdate(_ api: MEGAChatSdk, call: MEGAChatCall) {
         if call.status == .inProgress,
            call.callCompositionChange == .peerAdded,
            call.chatId == chatId,
