@@ -230,7 +230,7 @@ final class ChatRoomRepositoryTests: XCTestCase {
         return sut
     }
     
-    private func expectError<T>(from function: () async throws -> T, file: StaticString = #file, line: UInt = #line) async {
+    private func expectError<T>(from function: () async throws -> T, file: StaticString = #filePath, line: UInt = #line) async {
         do {
             _ = try await function()
             XCTFail("Expected an error to be thrown", file: file, line: line)
